Add the `sink` member to Request, which allows multiple messages to be send by a handler
This still requires more testing, but in theory it should allow handlers to
clone the `sink` and do _other_ asynchronous outbound writes to the websocket.
In essence, a handler could send dozens of messages out to the websocket when
handling the inbound message, rather than just a single message.

This still limits reads to the "normal" control flow in Meows